requirements.txt
python-geoip==1.2
python-geoip-geolite2==2014.207
